LayoutTests/editing/selection/user-select-none-02-expected.txt

You should not be able of select this text.
PASS selection.toString() is ""
PASS successfullyParsed is true

TEST COMPLETE
